refactor(frontend): split web chat screen responsibilities
