+++ apps/web/src/lib/domain/chat/cloud-chat-transcript-row-presentation.ts
@@ -1,0 +1,17 @@
+import type { CloudChatTranscriptRowView } from "@proliferate/product-domain/chats/cloud/transcript-view";
+
+export function isAssistantLoadingRow(row: CloudChatTranscriptRowView): boolean {
+  return row.kind === "assistant"
+    && Boolean(row.streaming)
+    && (
+      !row.body?.trim()
+      || row.id.includes(":assistant-waiting")
+      || row.id.includes(":pending-assistant")
+    );
+}
+
+export function loadingStatusText(row: CloudChatTranscriptRowView): string | null {
+  const value = row.detail ?? row.body ?? row.status ?? null;
+  const trimmed = value?.trim();
+  return trimmed ? trimmed : null;
+}
